updated framework to JUCE v1.53 (partial merge)

+++ src/audio_file_player.h
@@ -44,7 +44,7 @@ public:
 
 private:
     bool bIsPlaying;
-    int nNumberOfSamples;
+	int64 nNumberOfSamples;
     float fSampleRate;
 
     int nReportChannel;
